turned RevisionID into long
git-svn-id: https://svn.code.sf.net/p/autowikibrowser/code@4735 5227b50d-9349-4574-b055-4b0a7525e11c

+++ AWB/WikiFunctions/ErrorHandler.cs
@@ -26,7 +26,7 @@ namespace WikiFunctions
         /// <summary>
         /// Revision of the page currently being processed
         /// </summary>
-        public static int CurrentRevision;
+        public static long CurrentRevision;
 
         /// <summary>
         /// Current text that the list is being made from in ListMaker
